ERE-1812 Refactor Patient List column configuration * Create new class to define and load column definitions from registry metadata * Move Column classes to their own file for tidiness
